+++ Frontend/src/pages/Home.jsx
@@ -7,7 +7,7 @@ import Navbar from "../components/Navbar";
 import "./Home.css";
 import Loading from "../components/Loading";
 import { FaSearch } from "react-icons/fa";
-import toast from "react-hot-toast";
+import Swal from 'sweetalert2';
 
 
 const HomePage = () => {
@@ -49,55 +49,35 @@ const HomePage = () => {
   }, [selectedCategory]);
 
   // add or remove meal from favourites with confirmation using toast to look atractive
-  const toggleFavouriteMeal = (meal) => {
-    const alreadyLiked = favouriteMeals.find((item) => item.idMeal === meal.idMeal);
+ const toggleFavouriteMeal = (meal) => {
+  const alreadyLiked = favouriteMeals.find((item) => item.idMeal === meal.idMeal);
 
-    if (alreadyLiked) {
-      toast((t) => (
-        <span>
-          <b>{meal.strMeal}</b> is already in favourites.<br />
-          <button
-            style={{
-              background: "#f87171",
-              color: "#fff",
-              border: "none",
-              marginTop: "8px",
-              padding: "6px 12px",
-              borderRadius: "5px",
-              cursor: "pointer"
-            }}
-            onClick={() => {
-              const updatedList = favouriteMeals.filter((item) => item.idMeal !== meal.idMeal);
-              setFavouriteMeals(updatedList);
-              sessionStorage.setItem("favouriteMeals", JSON.stringify(updatedList));
-              toast.dismiss(t.id);
-              toast.success(`Removed "${meal.strMeal}" from favourites.`);
-            }}
-          >
-            Yes, Remove
-          </button>
-          <button
-            style={{
-              marginLeft: "10px",
-              marginTop: "8px",
-              padding: "6px 12px",
-              border: "1px solid #aaa",
-              borderRadius: "5px",
-              cursor: "pointer"
-            }}
-            onClick={() => toast.dismiss(t.id)}
-          >
-            Cancel
-          </button>
-        </span>
-      ), { duration: 5000 });
-    } else {
-      const updatedList = [...favouriteMeals, meal];
-      setFavouriteMeals(updatedList);
-      sessionStorage.setItem("favouriteMeals", JSON.stringify(updatedList));
-      toast.success(`Added "${meal.strMeal}" to favourites.`);
-    }
-  };
+  if (alreadyLiked) {
+    Swal.fire({
+      title: `"${meal.strMeal}" is already in favourites.`,
+      text: "Do you want to remove it?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#ef4444",
+      cancelButtonColor: "#6b7280",
+      confirmButtonText: "Yes, remove it"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const updatedList = favouriteMeals.filter((item) => item.idMeal !== meal.idMeal);
+        setFavouriteMeals(updatedList);
+        sessionStorage.setItem("favouriteMeals", JSON.stringify(updatedList));
+
+        Swal.fire("Removed!", `"${meal.strMeal}" was removed from favourites.`, "success");
+      }
+    });
+  } else {
+    const updatedList = [...favouriteMeals, meal];
+    setFavouriteMeals(updatedList);
+    sessionStorage.setItem("favouriteMeals", JSON.stringify(updatedList));
+
+    Swal.fire("Added!", `"${meal.strMeal}" has been added to favourites.`, "success");
+  }
+};
 
 
   // fetch detailed information of a meal by meal ID and update state
